Added _calculate_log_likelihood method and modified fit method.
The _calculate_log_likelihood method implements an ELBO based
lower bound estimate of the log likelihood of the observed data.
I wanted this function so that I could compare various hill climbs
of the same data to see which performed the best.

I also prefer using the estimate of the log likelihood as a halting
criteria than the L2 norm of the tau parameters.